updated join cohort page
